Added token for git push credentials

+++ usecases/rebuild_graph_usecase.py
@@ -10,12 +10,16 @@ from services.gtfs_service import GTFSService
 
 
 class RebuildGraphUseCase:
+
     def __init__(self, db, otp_db, user_id):
+        token = os.environ["GITHUB_TOKEN"]
         self.db = db
         self.otp_db = otp_db
         self.user_id = user_id
-        self.GITHUB_REPO = "https://github.com/Ruth652/taximela-backend.git"
-
+        self.GITHUB_REPO = (
+            f"https://{token}@github.com/Ruth652/taximela-backend.git"
+        )
+        
         # Local repo path (your machine for now)
         # self.deploy_repo_path = r"C:\TaxiMelaProject\taximela-otp-deploy"
         # self.branch = "otp-deploy-clean"
@@ -98,12 +102,21 @@ class RebuildGraphUseCase:
             shutil.rmtree(repo_dir)
             return
 
-        subprocess.run(
+        result = subprocess.run(
             ["git", "push", "origin", "otp-deploy-clean"],
             cwd=repo_dir,
-            check=True
+            capture_output=True,
+            text=True
         )
+        
+        print("STDOUT:\n", result.stdout)
+        print("STDERR:\n", result.stderr)
 
+        if result.returncode != 0:
+            raise Exception(
+                f"Git push failed\nSTDOUT:\n{result.stdout}\nSTDERR:\n{result.stderr}"
+            )
+        
         print("🚀 Pushed → Railway will auto-deploy")
 
         shutil.rmtree(repo_dir)
